[Network] Remove appended strings from the etag.
Some etags sent from the server include appended text, such as -gzip
when the request is gzipped. The etag spec requires that etags be
defined as a key inside of a pair of quotes, and nothing else.

With that in light, we find the quoted key in the etag and only cache
that value.

This has been tested with gzipped twitter requests and correctly
returns a 304 when used with just the key (sans -gzip suffix).