Top bar is always on for last screen.
There's a bug where it'll be on if you drag back from last. This
was really hard to fix because views are updated in a different flow
from when scroll view events are called. Also, if we do it purely in
view updates, we end up removing the bar one someone is using the left
right buttons as well.